+++ Sadie.Game.Rooms/RoomTile.cs
@@ -7,7 +7,7 @@ public class RoomTile
     public HPoint Point { get; }
     public RoomTileState State { get; }
     
-    public RoomTile(int x, int y, double z, RoomTileState state)
+    public RoomTile(int x, int y, float z, RoomTileState state)
     {
         Point = new HPoint(x, y, z);
         State = state;
